% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{check_tica}
\alias{check_tica}
\title{check_tica}
\usage{
check_tica(version = "0.2.2", stop_on_false = TRUE)
}
\arguments{
\item{version}{char with version number}

\item{stop_on_false}{bool}
}
\value{
silent
}
\description{
This functions check version of templateICAr library. If compares false, it stops if stop_on_false = TRUE
}
\details{
check version of templateICAr library, if not desired, choose if stop
}
\keyword{internal}
