% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{component_index.relevant}
\alias{component_index.relevant}
\title{relevant components index}
\usage{
component_index.relevant()
}
\arguments{
\item{NULL}{}
}
\value{
vector with RSN indices
}
\description{
This functions returns a vector with the names of the RSN components.
It should be modified according to the groupICA results of that specific run
}
\details{
Get components indices for the dhcp analysis (CHANGES WITH EACH run of gICA)
}
\keyword{internal}
