% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{component_names}
\alias{component_names}
\title{Definitions}
\usage{
component_names(ommit = TRUE, spaces = TRUE)
}
\arguments{
\item{NULL}{}
}
\value{
vector with RSN names
}
\description{
This functions returns a vector with the names of the RSN components.
It should be modified according to the groupICA results of that specific run
}
\details{
Get components names for the dhcp analysis (CHANGES WITH EACH run of gICA)
}
\keyword{internal}
