% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{default_mwall}
\alias{default_mwall}
\title{Medial wall}
\usage{
default_mwall(env)
}
\arguments{
\item{par}{This is a test}
}
\value{
The path to workbench command
}
\description{
This function asks for the hosts and returns the appropriate path to wb_command binary
}
\details{
Get dhcp midthickness symatalas medialwall
}
\keyword{internal}
