% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{dhcp_list}
\alias{dhcp_list}
\title{List of Subjects}
\usage{
dhcp_list(table_fname)
}
\arguments{
\item{table:}{input table}
}
\value{
list of lists with the duple (subid, sesid)
}
\description{
This function takes a table of subjects in csv format and returns a list of (subid, sesid) tuples(lists)
}
\details{
Get subid and sesid of a set of subjects, given a csv. TODO: Add filtering by conditions?
}
\keyword{internal}
