% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{dhcp_path}
\alias{dhcp_path}
\title{DHCP path}
\usage{
dhcp_path()
}
\arguments{
\item{par}{This is a test}
}
\value{
The path to workbench command
}
\description{
This function asks for the hosts and returns the appropriate path to wb_command binary
}
\details{
Get dhcp path depending on the host
}
\keyword{internal}
