% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{tmap}
\alias{tmap}
\title{utils.r}
\usage{
tmap(
  subid,
  sesid,
  template_name,
  save_png = FALSE,
  save_rds = NULL,
  method_p = "fdr",
  alpha = 0.05,
  env = NULL
)
}
\arguments{
\item{subid}{DHCP subject id number}

\item{sesid}{DHCP session id number}

\item{template_name}{DHCP template name of the input dual regression estimates}

\item{save_png}{toggle to save tstat and binary images. Default: \code{FALSE}}

\item{save_rds}{path to save RDS file out of \code{templateICAr::activations()}}

\item{method_p}{multiple comparison p correction method from \code{templateICAr::activations()}}

\item{alpha}{threshold value for one-sample t-test}

\item{env}{an environment with the default surfaces, to save time.
If null, will load dhcp symatlas surf.}
}
\value{
Void, save images if set.
}
\description{
Utility functions for the dhcp dataset
}
\details{
Make thresholded t-statistic maps.

Performs one-sample t-test and writes tmap rds, cifti, and png for reporting
}
