% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.r
\name{V}
\alias{V}
\title{Workbench Paths and Definitions}
\usage{
V()
}
\arguments{
\item{par}{This is a test}
}
\value{
The path to workbench command
}
\description{
This function asks for the hosts and returns the appropriate path to wb_command binary
}
\details{
Get workbench path depending on the host
}
\keyword{internal}
